User-Management-System/Backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-decouple==3.8
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
